pyproject.toml
[tool.poetry]
name = "pipedrive_metabase_integration"
version = "0.1.0"
description = "Integração do Pipedrive com Metabase utilizando Airflow"
authors = ["Pavcob"]
readme = ["README.md"]

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
prefect = "^3.2.14"
requests = "2.32.3"
python-decouple = "3.8"
python-dotenv = "1.0.1"
pydantic = '2.10.6'
prometheus-client = "0.21.1"
psycopg2  = '2.9.10'
httpx='0.28.1'
tenacity='9.0.0'
apache_beam='2.63.0'
python-json-logger='3.3.0'

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
